Create new desktop files with all required content
the html has all the custom fonts linked.

html consists of sidebar section and signup section. The form is
included in the signup section and the form controls are split into
groups of two inputs, using sections. The submit button, in the signup
section is not inside the form but is linked to the form using the
form attribute.

Applied a CSS reset to signup.css and also added sidebar image as
background.