Apply flexbox to sidebar and fix width issues
Applied flexbox to sidebar to organize logo and background attribution.

Set a min width on the sidebar because I did not want the text content
of the sidebar to wrap, mainly because the logo needs all text in one
line.

Realized that the content of the container was overflowing
horizontally and so set a min width on the .container